# src/main/resources/openapi.yml
openapi: 3.1.0
info:
  title: User Management Service API
  description: RESTful API for users management
  contact:
    name: Umbarry
    email: umberto@atania.eu
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearer-jwt: []
paths:
  /v1/users/{id}:
    get:
      tags:
        - user-controller
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    put:
      tags:
        - user-controller
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    delete:
      tags:
        - user-controller
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
  /v1/users/{id}/status:
    put:
      tags:
        - user-controller
      operationId: updateUserStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
        required: true
      responses:
        "200":
          description: OK
  /v1/users:
    get:
      tags:
        - user-controller
      operationId: getAllUsers
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
    post:
      tags:
        - user-controller
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
components:
  schemas:
    UserRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        taxCode:
          type: string
          minLength: 1
          pattern: "^[A-Z]{6}\\d{2}[A-Z]\\d{2}[A-Z]\\d{3}[A-Z]$"
        name:
          type: string
          minLength: 1
        surname:
          type: string
          minLength: 1
        status:
          type: string
          enum:
            - ACTIVE
            - DISABLED
            - DELETED
        roles:
          type: array
          items:
            type: string
            enum:
              - OWNER
              - OPERATOR
              - MAINTAINER
              - DEVELOPER
              - REPORTER
          minItems: 1
          uniqueItems: true
      required:
        - email
        - name
        - roles
        - status
        - surname
        - taxCode
        - username
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        taxCode:
          type: string
        name:
          type: string
        surname:
          type: string
        status:
          type: string
          enum:
            - ACTIVE
            - DISABLED
            - DELETED
        roles:
          type: array
          items:
            type: string
            enum:
              - OWNER
              - OPERATOR
              - MAINTAINER
              - DEVELOPER
              - REPORTER
          uniqueItems: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum:
            - ACTIVE
            - DISABLED
            - DELETED
          minLength: 1
      required:
        - status
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
  securitySchemes:
    bearer-jwt:
      type: http
      description: JWT authentication token from Keycloak
      scheme: bearer
      bearerFormat: JWT